Fixed dtype typo for int64 insert

+++ pandas/tests/indexing/test_coercion.py
@@ -415,7 +415,7 @@ class TestInsertIndexCoercion(CoercionBase):
     @pytest.mark.parametrize("insert, coerced_val, coerced_dtype", [
         (1, 1, np.int64),
         (1.1, 1.1, np.float64),
-        (False, 0, int),
+        (False, 0, np.int64),
         ('x', 'x', np.object)])
     def test_insert_index_int64(self, insert, coerced_val, coerced_dtype):
         obj = pd.Int64Index([1, 2, 3, 4])
